+++ python_ta/checkers/invalid_for_target_checker.py
@@ -1,4 +1,4 @@
-"""Checker for target of for loop in subscript form.
+"""Checker for target of for loop or comprehension in subscript form.
 """
 from typing import List, Union
 
@@ -13,9 +13,9 @@ class InvalidForTargetChecker(BaseChecker):
     # use dashes for connecting words in message symbol
     msgs = {
         "E9984": (
-            'For loop variable "%s" should not be a part of a larger object.',
+            'For loop or comprehension variable "%s" should not be a part of a larger object.',
             "invalid-for-target",
-            "Used when you have a loop variable in a for loop "
+            "Used when you have an index variable in a for loop or comprehension"
             "that is in subscript or object attribute form",
         )
     }
@@ -26,6 +26,12 @@ class InvalidForTargetChecker(BaseChecker):
 
     @only_required_for_messages("invalid-for-target")
     def visit_for(self, node: nodes.For) -> None:
+        invalid_for_targets = node.target.nodes_of_class(self.INVALID_TARGETS)
+        for target in invalid_for_targets:
+            self.add_message("invalid-for-target", node=target, args=target.as_string())
+
+    @only_required_for_messages("invalid-for-target")
+    def visit_comprehension(self, node: nodes.Comprehension) -> None:
         invalid_for_targets = node.target.nodes_of_class(self.INVALID_TARGETS)
         for target in invalid_for_targets:
             self.add_message("invalid-for-target", node=target, args=target.as_string())