+++ python_ta/contracts/__init__.py
@@ -184,7 +184,7 @@ def _check_function_contracts(wrapped, instance, args, kwargs):
         if param in annotations:
             try:
                 _debug(f"Checking type of parameter {param} in call to {wrapped.__qualname__}")
-                check_type(param, arg, annotations[param])
+                check_type_strict(param, arg, annotations[param])
             except TypeError:
                 additional_suggestions = _get_argument_suggestions(arg, annotations[param])
 
@@ -224,7 +224,7 @@ def _check_function_contracts(wrapped, instance, args, kwargs):
         return_type = annotations["return"]
         try:
             _debug(f"Checking return type from call to {wrapped.__qualname__}")
-            check_type("return", r, return_type)
+            check_type_strict("return", r, return_type)
         except TypeError:
             raise PyTAContractError(
                 f"{wrapped.__name__}'s return value {_display_value(r)} did not match "
@@ -257,6 +257,13 @@ def _check_function_contracts(wrapped, instance, args, kwargs):
     )
 
     return r
+
+
+def check_type_strict(argname: str, value: Any, expected_type: type) -> None:
+    """Ensure that ``value`` matches ``expected_type``. Differentiates between types float and int."""
+    if type(value) is int and expected_type is float:
+        raise TypeError(f"type of {argname} must be {expected_type}; got {value} instead")
+    check_type(argname, value, expected_type)
 
 
 def _get_argument_suggestions(arg: Any, annotation: type) -> str: