+++ python_ta/contracts/__init__.py
@@ -18,6 +18,7 @@ from typing import Any, Callable, List, Optional, Set, Tuple
 import wrapt
 from typeguard import check_type
 
+ENABLE_CONTRACT_CHECKING = True
 DEBUG_CONTRACTS = False
 """
 Set to True to display debugging messages when checking contracts.
@@ -157,7 +158,7 @@ def add_class_invariants(klass: type) -> None:
         if self is not frame_locals.get("self"):
             # Only validating if the attribute is not being set in a instance/class method
             klass_mod = sys.modules.get(klass.__module__)
-            if klass_mod is not None:
+            if klass_mod is not None and ENABLE_CONTRACT_CHECKING:
                 try:
                     _check_invariants(self, klass, klass_mod.__dict__)
                 except PyTAContractError as e:
@@ -216,7 +217,8 @@ def _check_function_contracts(wrapped, instance, args, kwargs):
                 continue
             target.__preconditions__.append((precondition, compiled))
 
-    _check_assertions(wrapped, function_locals)
+    if ENABLE_CONTRACT_CHECKING:
+        _check_assertions(wrapped, function_locals)
 
     # Check return type
     r = wrapped(*args, **kwargs)
@@ -293,7 +295,7 @@ def _instance_method_wrapper(wrapped: Callable, klass: type) -> Callable:
                 return r
             _check_class_type_annotations(klass, instance)
             klass_mod = sys.modules.get(klass.__module__)
-            if klass_mod is not None:
+            if klass_mod is not None and ENABLE_CONTRACT_CHECKING:
                 _check_invariants(instance, klass, klass_mod.__dict__)
         except PyTAContractError as e:
             raise AssertionError(str(e)) from None