+++ tests/test_contracts.py
@@ -3,6 +3,7 @@ from typing import Dict, List, Set
 
 import pytest
 
+import python_ta.contracts
 from python_ta.contracts import check_contracts
 
 
@@ -122,6 +123,19 @@ def test_nullary_return_bool_int_error() -> None:
         nullary()
 
 
+def test_nullary_int_bool_disable_contract_checking() -> None:
+    """Calling a nullary function with incorrect return type and with ENABLE_CONTRACT_CHECKING disabled so no error
+    is raised."""
+
+    @check_contracts
+    def nullary() -> int:
+        return True
+
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = False
+    nullary()
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = True  # Reset default value to True
+
+
 def test_nullary_no_return_type() -> None:
     """Calling a nullary function with no specified return type passes."""
 
@@ -219,6 +233,16 @@ def test_parameter_bool_int_error() -> None:
         parameter_bool(1)
 
 
+def test_parameter_int_bool_disable_contract_checking() -> None:
+    @check_contracts
+    def parameter_int(num: int) -> None:
+        return None
+
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = False
+    parameter_int(True)
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = True  # Reset default value to True
+
+
 @check_contracts
 def _my_sum_one_precondition(numbers: List[int]) -> int:
     """Return the sum of a list of numbers.
@@ -248,6 +272,14 @@ def test_my_sum_one_pre_violation() -> None:
 
     msg = str(excinfo.value)
     assert "len(numbers) > 2" in msg
+
+
+def test_my_sum_one_disable_contract_checking() -> None:
+    """Calling _my_sum_one_precondition with a value that violates the precondition but with ENABLE_CONTRACT_CHECKING
+    = False so no error is raised"""
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = False
+    _my_sum_one_precondition([1])
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = True  # Reset default value to True
 
 
 # Checking to see if functions we defined are in-scope for preconditions
@@ -415,6 +447,15 @@ def test_get_double_invalid() -> None:
 
     msg = str(excinfo.value)
     assert "$return_value == num * 2" in msg
+
+
+def test_get_double_disabled_contract_checking() -> None:
+    """Test that calling the invalid implementation of _get_double does NOT raise an AssertionError when
+    ENABLE_CONTRACT_CHECKING is False.
+    """
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = False
+    assert _get_double_invalid(5) == 11
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = True  # Reset default value to True
 
 
 # Test that postcondition checks involving function parameters pass and fail as expected