+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `forbidden_io_function_checker` is now able to check for calls to IO functions written at the top-level of a module, but outside the main block.
 - `python_ta.debug.AccumulationTable` is extended to support printing loop iterations for while loops.
 - Violated representation invariant error message now includes the class name and current values of the instance attributes.
+- Added constant `ENABLE_CONTRACT_CHECKING` to only check preconditions/representation invariants when its value is set to `True`.
 
 ### Bug Fixes